requirements.txt
numpy
opencv-python-headless
streamlit
mysql-connector-python
pandas
tensorflow
face-recognition